# Conflicts:
#
fengshun/fengshun.xcodeproj/project.xcworkspace/xcuserdata/aiwei.xcuserd
atad/UserInterfaceState.xcuserstate
#
fengshun/fengshun.xcodeproj/xcuserdata/aiwei.xcuserdatad/xcdebugger/Brea
kpoints_v2.xcbkptlist
#	fengshun/fengshun/mainTab/FSMainVC.m

+++ fengshun/fengshun/AppDelegate.m
@@ -80,7 +80,7 @@
 - (void)initILiveSDK
 {
     // 初始化SDK
-    [[ILiveSDK getInstance] initSdk:FS_ILiveSDKAPPID accountType:FS_ILiveAccountType];
+    [[ILiveSDK getInstance] initSdk:(int)FS_ILiveSDKAPPID accountType:(int)FS_ILiveAccountType];
     [[ILiveSDK getInstance] setChannelMode:E_ChannelIMSDK withHost:@""];
     
     // 获取版本号
